+++ CHANGELOG.md
@@ -215,6 +215,7 @@ Initial beta release !
 - New artifacts and creatures from the mod added to the editor
 - Added spells and training bonuses sections in creatures description window
 - More artifacts (Ardent Smith, Genji's set and many that were disabled for now working)
+- Changed map objects effects (Idol of Fortune, Tomb of the Warrior, Seer Hut, War Academy, Arcane Sanctuary, Mother Earth Shrine)
 
 #### Fixed/Changed